fix local MSP id for orderer
local MSP ID for orderer was CaseOrdererORG instead of CaseOrdererMSP. Due to this, it was failing to create the channel. The following error was thrown:
"'implicit policy evaluation failed - 0 sub-policies were satisfied, but this policy requires 1 of the \'Writers\' sub-policies to be satisfied: permission denied'"

Upon further investigation, the following error was thrown in network.log:
"Principal deserialization failure (MSP CaseOrdererOrg is unknown) for identity 0"

Upon changing this, the issue was resolved.